add DAGRequest
