+++ dbms/src/Flash/Coprocessor/DAGContext.cpp
@@ -18,7 +18,6 @@
 #include <Flash/Coprocessor/RequestUtils.h>
 #include <Flash/Coprocessor/collectOutputFieldTypes.h>
 #include <Flash/Mpp/ExchangeReceiver.h>
-#include <Flash/Planner/ExecutorIdGenerator.h>
 #include <Flash/Statistics/traverseExecutors.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <tipb/executor.pb.h>
@@ -57,15 +56,7 @@ DAGContext::DAGContext(tipb::DAGRequest & dag_request_, TablesRegionsInfo && tab
     , warning_count(0)
     , keyspace_id(keyspace_id_)
 {
-    RUNTIME_CHECK((dag_request->executors_size() > 0) != isTreeBasedExecutors());
-    const auto & root_executor = isTreeBasedExecutors()
-        ? dag_request->root_executor()
-        : dag_request->executors(dag_request->executors_size() - 1);
-    return_executor_id = root_executor.has_executor_id();
-    if (return_executor_id)
-        root_executor_id = root_executor.executor_id();
     initOutputInfo();
-    initListBasedExecutors();
 }
 
 // for mpp
@@ -74,7 +65,6 @@ DAGContext::DAGContext(tipb::DAGRequest & dag_request_, const mpp::TaskMeta & me
     , dummy_query_string(dag_request->DebugString())
     , dummy_ast(makeDummyQuery())
     , collect_execution_summaries(dag_request->has_collect_execution_summaries() && dag_request->collect_execution_summaries())
-    , return_executor_id(true)
     , is_mpp_task(true)
     , is_root_mpp_task(is_root_mpp_task_)
     , flags(dag_request->flags())
@@ -86,12 +76,9 @@ DAGContext::DAGContext(tipb::DAGRequest & dag_request_, const mpp::TaskMeta & me
     , warning_count(0)
     , keyspace_id(RequestUtils::deriveKeyspaceID(meta_))
 {
-    RUNTIME_CHECK(isTreeBasedExecutors() && dag_request->root_executor().has_executor_id());
-    root_executor_id = dag_request->root_executor().executor_id();
     // only mpp task has join executor.
     initExecutorIdToJoinIdMap();
     initOutputInfo();
-    initListBasedExecutors();
 }
 
 DAGContext::DAGContext(tipb::DAGRequest & dag_request_, const DM::DisaggTaskId & task_id_, TablesRegionsInfo && tables_regions_info_, const String & compute_node_host_, LoggerPtr log_)
@@ -113,11 +100,7 @@ DAGContext::DAGContext(tipb::DAGRequest & dag_request_, const DM::DisaggTaskId &
     , warnings(max_recorded_error_count)
     , warning_count(0)
 {
-    RUNTIME_CHECK(isTreeBasedExecutors() && dag_request->root_executor().has_executor_id());
-    return_executor_id = dag_request->root_executor().has_executor_id() || dag_request->executors(0).has_executor_id();
-
     initOutputInfo();
-    initListBasedExecutors();
 }
 
 // for test
@@ -150,15 +133,7 @@ DAGContext::DAGContext(tipb::DAGRequest & dag_request_, String log_identifier, s
     , warnings(max_recorded_error_count)
     , warning_count(0)
 {
-    RUNTIME_CHECK((dag_request->executors_size() > 0) != isTreeBasedExecutors());
-    const auto & root_executor = isTreeBasedExecutors()
-        ? dag_request->root_executor()
-        : dag_request->executors(dag_request->executors_size() - 1);
-    return_executor_id = root_executor.has_executor_id();
-    if (return_executor_id)
-        root_executor_id = root_executor.executor_id();
     initOutputInfo();
-    initListBasedExecutors();
 }
 
 void DAGContext::initOutputInfo()
@@ -177,29 +152,6 @@ void DAGContext::initOutputInfo()
     }
     encode_type = analyzeDAGEncodeType(*this);
     keep_session_timezone_info = encode_type == tipb::EncodeType::TypeChunk || encode_type == tipb::EncodeType::TypeCHBlock;
-}
-
-
-void DAGContext::initListBasedExecutors()
-{
-    if (!isTreeBasedExecutors())
-    {
-        ExecutorIdGenerator id_generator;
-        for (int i = 0; i < dag_request->executors_size(); ++i)
-        {
-            auto * executor = dag_request->mutable_executors(i);
-            const auto & executor_id = id_generator.generate(*executor);
-            list_based_executors_order.push_back(executor_id);
-            // Set executor_id for list based executor,
-            // then we can fill executor_id for Execution Summaries of list-based executors
-            executor->set_executor_id(executor_id);
-        }
-    }
-}
-
-bool DAGContext::isTreeBasedExecutors() const
-{
-    return dag_request->has_root_executor();
 }
 
 bool DAGContext::allowZeroInDate() const
@@ -237,7 +189,7 @@ void DAGContext::initExecutorIdToJoinIdMap()
         return;
 
     executor_id_to_join_id_map.clear();
-    traverseExecutorsReverse(dag_request, [&](const tipb::Executor & executor) {
+    traverseExecutorsReverse(dag_request(), [&](const tipb::Executor & executor) {
         std::vector<String> all_join_id;
         // for mpp, dag_request.has_root_executor() == true, can call `getChildren` directly.
         getChildren(executor).forEach([&](const tipb::Executor & child) {
