+++ dbms/src/Flash/Statistics/ExecutionSummaryHelper.cpp
@@ -32,7 +32,7 @@ void fillTiExecutionSummary(
 
     // tree-based executors will have executor_id.
     // In ut, list-based executor will have executor_id for result comparision.
-    if (dag_context.return_executor_id || force_fill_executor_id)
+    if (dag_context.dag_request.isTreeBased() || force_fill_executor_id)
         execution_summary->set_executor_id(executor_id);
 }
 } // namespace DB
